- Separated guessing scans from mining scans - HeaderMutationScan -> HeaderSmugglingMutationScan
