+++ src/burp/model/scanning/guessing/param/LengthCompare.java
@@ -1,4 +1,4 @@
-package burp.model.param;
+package burp.model.scanning.guessing.param;
 
 import java.util.Comparator;
